merge处理
