+++ .gitignore
@@ -1,5 +1,5 @@
 /app/torrent
-/app/config/config.php
+/config/config.php
 /php-7.2.12-nts
 /*.bat
 /*.sh
