--- wiki/数据清理.md
@@ -1,21 +1,0 @@
-## 404数据最后清理时间
-|标志 | 最后清理时间 | 状态 |
-| :-: | :-: | ---- |
-|  ttg   |  |  |
-|城市 |  |  |
-|萌猫 | 2020年1月16日12:07:12 | 已完成 |
-|ptmsg | 2020年1月16日12:07:17 | 已完成 |
-|瓷器 |  |  |
-|杜比 | 2020年1月16日11:34:19 | 已完成 |
-|tccf |  |  |
-|馒头 |  |  |
-|春天 | 2020年1月16日00:23:05 | 已完成 |
-|北洋 | 2020年1月16日00:23:05 | 已完成 |
-|猫站 |  |  |
-|朋友 | 2020年1月16日00:23:05 | 已完成 |
-|铂金家 | 2020年1月16日00:23:05 | 已完成 |
-|家园 | 2020年1月16日00:23:05 | 已完成 |
-|学校 | 2020年1月16日00:19:24 | 已完成 |
-| 我堡 | 2020年1月15日13:56:22 | 已完成 |
-| 天空 | 2020年1月15日20:49:33 | 已完成 |
-
